requirements.txt
python-dotenv
jsonargparse
langchain
langchain-community
openai
moviepy